Add files via upload
Code based on the merge sort algorithm code from https://hackr.io/blog/python-projects#intermediate-python-projects-with-source-code with adjustments by myself. When prompted, enter an integer representing the length you want the list to be sorted to be. Then, enter a series of integers, hitting enter after each, until the list is full. Then it will automatically sort the list of integers until it is in ascending order.